refactor: moved AbilitySource to it's own package, split individual sources into their own classes
