+++ src/main/java/com/wanderersoftherift/wotr/abilities/StoredAbilityContext.java
@@ -2,6 +2,7 @@ package com.wanderersoftherift.wotr.abilities;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import com.wanderersoftherift.wotr.abilities.sources.AbilitySource;
 import net.minecraft.core.Holder;
 import net.minecraft.core.UUIDUtil;
 import net.minecraft.server.MinecraftServer;
