+++ src/main/java/com/wanderersoftherift/wotr/abilities/attachment/AbilityTracker.java
@@ -3,8 +3,8 @@ package com.wanderersoftherift.wotr.abilities.attachment;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import com.wanderersoftherift.wotr.abilities.Ability;
-import com.wanderersoftherift.wotr.abilities.AbilitySource;
 import com.wanderersoftherift.wotr.abilities.TrackedAbilityTrigger;
+import com.wanderersoftherift.wotr.abilities.sources.AbilitySource;
 import com.wanderersoftherift.wotr.init.WotrAttachments;
 import com.wanderersoftherift.wotr.init.WotrRegistries;
 import com.wanderersoftherift.wotr.item.ability.AbilityModifier;
