+++ src/main/java/com/wanderersoftherift/wotr/init/WotrAbilitySourceTypes.java
@@ -1,7 +1,9 @@
 package com.wanderersoftherift.wotr.init;
 
 import com.wanderersoftherift.wotr.WanderersOfTheRift;
-import com.wanderersoftherift.wotr.abilities.AbilitySource;
+import com.wanderersoftherift.wotr.abilities.sources.AbilitySource;
+import com.wanderersoftherift.wotr.abilities.sources.MainAbilitySource;
+import com.wanderersoftherift.wotr.abilities.sources.ModifierAbilitySource;
 import com.wanderersoftherift.wotr.serialization.DualCodec;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
@@ -12,7 +14,7 @@ public class WotrAbilitySourceTypes {
             .create(WotrRegistries.ABILITY_SOURCES, WanderersOfTheRift.MODID);
 
     public static final Supplier<DualCodec<? extends AbilitySource>> MODIFIER_SOURCE = ABILITY_SOURCE
-            .register("modifier", () -> AbilitySource.ModifierAbilitySource.TYPE);
+            .register("modifier", () -> ModifierAbilitySource.TYPE);
     public static final Supplier<DualCodec<? extends AbilitySource>> MAIN_SOURCE = ABILITY_SOURCE.register("main",
-            () -> AbilitySource.MainAbilitySource.TYPE);
+            () -> MainAbilitySource.TYPE);
 }
