+++ src/main/java/com/wanderersoftherift/wotr/modifier/source/AbilityUpgradeModifierSource.java
@@ -2,7 +2,7 @@ package com.wanderersoftherift.wotr.modifier.source;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.wanderersoftherift.wotr.abilities.AbilitySource;
+import com.wanderersoftherift.wotr.abilities.sources.AbilitySource;
 import com.wanderersoftherift.wotr.modifier.effect.AbstractModifierEffect;
 import com.wanderersoftherift.wotr.serialization.DualCodec;
 import net.minecraft.network.codec.ByteBufCodecs;
